practice/CASE-三国演义Embedding/pyproject.toml
[project]
name = "case-sanguoyanyi-embedding"
version = "0.1.0"
description = "三国演义文本 Embedding 和向量检索项目"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0,<2.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "jieba>=0.42.0",
    "gensim>=4.3.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
]